requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.3.4
pypdf==6.10.2
python-multipart==0.0.27
sqlalchemy>=2.0.0
anthropic>=0.100.0
apscheduler>=3.11.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
python-docx>=1.2.0
